cleanup react code
