+++ src/game/index.html
@@ -105,6 +105,6 @@
         You have 3 lives. Complete all levels to win!
       </p>
     </div>
-    <script type="module" src="./main.ts"></script>
+    <script type="module" src="./main.tsx"></script>
   </body>
 </html>
